requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==1.5.3
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.0